Deleted old code and moved around types
